+++ app/state/[state]/session/[session]/bill/[bill]/votes/page.tsx
@@ -28,7 +28,7 @@ export default async function Page({
 
   return (
     <>
-      <div className="flex justify-center space-x-12 mt-6 text-lg">
+      <div className="flex justify-center space-x-12 mb-10 text-lg">
         <Link href={`/state/${params.state}/session/${params.session}/bill/${bill.bill_id}/history`}>
           <p className="cursor-pointer font-semibold">History</p>
         </Link>
@@ -58,7 +58,7 @@ export default async function Page({
             </li>
           ))
         ) : (
-          <li className="col-span-full text-center">No votes.</li>
+          <li className="border rounded-lg p-4 shadow-sm bg-white">No votes.</li>
         )}
       </ul>
     </>
