+++ data/apiResponse.ts
@@ -1,9 +1,3 @@
-// import MasterList from "./masterList";
-
-// export default interface ApiResponse {
-//     status: string;
-//     masterlist: MasterList;
-// }
 import SearchResult from "./searchResult";
 
 export default interface ApiResponse {
